Move globalAll into all using isBase() check
To make the product API more consistent for all method (and also future
methods), isBase() method was created to discern the original Product
class from children. This distinction is crucial in some parts where a
static method makes sense only on children classes but not on Product
itself (like old version of `all()`).