Index page logic (#3)
* Make and display ProductDisc index cards

* Return all objects in Api::index()

For now the view rendering logic has repetition and the product class
list is hardcoded.

* Sort api::index() results by id

* Pass data instead of template via api

This is done to make styling more flexible on frontend.

* Init redux

* Implement mass-deletion frontend

* Finish frontend logic and add mass-delete api func

* Move globalAll into all using isBase() check

To make the product API more consistent for all method (and also future
methods), isBase() method was created to discern the original Product
class from children. This distinction is crucial in some parts where a
static method makes sense only on children classes but not on Product
itself (like old version of `all()`).

* Create base fromId method

* Add deletion query and finish mass_delete api

* Abstract `indexCardData` attributes and fix price

To get rid of redundancy, a parent method `indexCardData()` was made. It
returns data common to all products.

Also, before this commit, the price representation was broken and now it
is fixed.

* Add missing html ids